requirements.txt
streamlit
google-genai
pydantic
google-auth
google-auth-oauthlib
google-api-python-client
pillow
google-cloud-logging
